deployment/app/requirements.txt
streamlit~=1.36.0
pandas~=2.0.0
numpy~=1.25.0
scikit-learn~=1.5.0
plotly~=5.17.0
millify~=0.1.1

google-cloud-pubsub~=2.22.0